+++ src/main/java/net/silentchaos512/gear/traits/TraitSerializers.java
@@ -25,6 +25,7 @@ public final class TraitSerializers {
         register(NBTTrait.SERIALIZER);
         register(PotionEffectTrait.SERIALIZER);
         register(StatModifierTrait.SERIALIZER);
+        register(AttributeTrait.SERIALIZER);
     }
 
     private TraitSerializers() {}